+++ src/components/InputTable.jsx
@@ -3,10 +3,13 @@ import PropTypes from 'prop-types';
 import InputTableHeader from './InputTableHeader';
 import InputTableRow from './InputTableRow';
 import styles from '../../styles/cspace-input/InputTable.css';
+import labelToLegend from '../helpers/labelToLegend';
 
 const propTypes = {
   children: PropTypes.node,
   embedded: PropTypes.bool,
+  id: PropTypes.string,
+  label: PropTypes.node,
   renderLabel: PropTypes.func,
   renderAriaLabel: PropTypes.func,
 };
@@ -14,6 +17,8 @@ const propTypes = {
 const defaultProps = {
   children: undefined,
   embedded: undefined,
+  id: undefined,
+  label: undefined,
   renderLabel: undefined,
   renderAriaLabel: undefined,
 };
@@ -22,12 +27,17 @@ export default function InputTable(props) {
   const {
     children,
     embedded,
+    id,
+    label,
     renderLabel,
     renderAriaLabel,
   } = props;
 
+  const legend = labelToLegend(label);
+
   return (
-    <div className={styles.common}>
+    <fieldset className={styles.common} id={id}>
+      {legend}
       <InputTableHeader embedded={embedded} renderLabel={renderLabel}>
         {children}
       </InputTableHeader>
@@ -35,9 +45,10 @@ export default function InputTable(props) {
       <InputTableRow embedded={embedded} renderAriaLabel={renderAriaLabel}>
         {children}
       </InputTableRow>
-    </div>
+    </fieldset>
   );
 }
 
 InputTable.propTypes = propTypes;
 InputTable.defaultProps = defaultProps;
+InputTable.useLegend = true;